+++ game/player.py
@@ -1,4 +1,5 @@
 from game.tiles import BagTiles
+from game.tiles import Tile
 
 class Player:
 
@@ -9,11 +10,11 @@ class Player:
     def rellenar(self):
         self.tiles += self.bag_tiles.take(7 - len(self.tiles))
 
-    def has_letters(self, tiles):
-        tiles_counts = {tile.letter: tile.value for tile in tiles}
-        player_tiles_counts = {tile.letter: tile.value for tile in self.tiles}
+    def has_letters(self, word):
+        word_tiles_counts = {letter: word.count(letter) for letter in word}
+        player_tiles_counts = {tile.letter: self.tiles.count(tile) for tile in self.tiles}
 
-        for letter, count in tiles_counts.items():
+        for letter, count in word_tiles_counts.items():
             if letter not in player_tiles_counts or count > player_tiles_counts[letter]:
                 return False
         return True